Move the monitoring modules into a separate finatra-monitoring package
