+++ build.sbt
@@ -62,10 +62,15 @@ lazy val common_storage = doSharedLibrarySetup(project, "sbt_common/storage")
 lazy val finatra_messaging = doSharedLibrarySetup(project, "sbt_common/finatra-messaging")
   .dependsOn(common_messaging % "compile->compile;test->test")
   .dependsOn(finatra_storage % "compile->compile;test->test")
+  .dependsOn(finatra_monitoring % "compile->compile;test->test")
   .settings(libraryDependencies ++= Dependencies.finatraDependencies)
 
 lazy val finatra_storage = doSharedLibrarySetup(project, "sbt_common/finatra_storage")
   .dependsOn(common_storage % "compile->compile;test->test")
+  .settings(libraryDependencies ++= Dependencies.finatraDependencies)
+
+lazy val finatra_monitoring = doSharedLibrarySetup(project, "sbt_common/finatra-monitoring")
+  .dependsOn(common_monitoring % "compile->compile;test->test")
   .settings(libraryDependencies ++= Dependencies.finatraDependencies)
 
 lazy val api = doServiceSetup(project, "catalogue_api/api")
