+++ catalogue_pipeline/transformer/src/main/scala/uk/ac/wellcome/platform/transformer/Server.scala
@@ -16,8 +16,8 @@ import uk.ac.wellcome.finatra.messaging.{
   SQSClientModule,
   SQSConfigModule
 }
+import uk.ac.wellcome.finatra.monitoring.MetricsSenderModule
 import uk.ac.wellcome.finatra.storage.{S3ClientModule, S3ConfigModule}
-import uk.ac.wellcome.monitoring.MetricsSenderModule
 import uk.ac.wellcome.platform.transformer.modules.{TransformerWorkerModule, _}
 
 object ServerMain extends Server
