+++ data_api/snapshot_generator/src/main/scala/uk/ac/wellcome/platform/snapshot_generator/Server.scala
@@ -17,8 +17,8 @@ import uk.ac.wellcome.finatra.messaging.{
   SQSConfigModule
 }
 import uk.ac.wellcome.finatra.modules._
+import uk.ac.wellcome.finatra.monitoring.MetricsSenderModule
 import uk.ac.wellcome.finatra.storage.S3ConfigModule
-import uk.ac.wellcome.monitoring.MetricsSenderModule
 import uk.ac.wellcome.platform.snapshot_generator.finatra.SnapshotGeneratorModule
 import uk.ac.wellcome.platform.snapshot_generator.finatra.modules.{
   AkkaS3ClientModule,
