+++ sbt_common/finatra-monitoring/src/main/scala/uk/ac/wellcome/finatra/monitoring/MetricsSenderModule.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.monitoring
+package uk.ac.wellcome.finatra.monitoring
 
 import akka.actor.ActorSystem
 import com.amazonaws.services.cloudwatch.AmazonCloudWatch
@@ -6,6 +6,7 @@ import com.google.inject.{Provides, Singleton}
 import com.twitter.app.Flaggable
 import com.twitter.inject.TwitterModule
 import uk.ac.wellcome.finatra.modules.AkkaModule
+import uk.ac.wellcome.monitoring.MetricsSender
 
 import scala.concurrent.duration._
