+++ sbt_common/monitoring/src/main/scala/uk/ac/wellcome/monitoring/CloudWatchClientFactory.scala
@@ -1,0 +1,21 @@
+package uk.ac.wellcome.monitoring
+
+import com.amazonaws.client.builder.AwsClientBuilder.EndpointConfiguration
+import com.amazonaws.services.cloudwatch.{
+  AmazonCloudWatch,
+  AmazonCloudWatchClientBuilder
+}
+
+object CloudWatchClientFactory {
+  def createClient(region: String, endpoint: String): AmazonCloudWatch = {
+    val standardClient = AmazonCloudWatchClientBuilder.standard
+    if (endpoint.isEmpty)
+      standardClient
+        .withRegion(region)
+        .build()
+    else
+      standardClient
+        .withEndpointConfiguration(new EndpointConfiguration(endpoint, region))
+        .build()
+  }
+}
